GHSA-pcvr-f2f2-xq7m: add Desktop database encryption
- refactor Database factory code
- use cipher for Desktop FFI as well
- build SQfLite factory with sqlcipher support
