+++ lib/utils/matrix_sdk_extensions/flutter_matrix_dart_sdk_database/sqlite_ffi/stub.dart
@@ -1,0 +1,4 @@
+/// overrides the sqlite shared object / dynamic library with the SQLCipher one
+///
+/// https://github.com/tekartik/sqflite/blob/master/sqflite_common_ffi/doc/encryption_support.md
+void ffiInit() {}
