+++ pubspec.yaml
@@ -81,6 +81,7 @@ dependencies:
   sqflite: ^2.3.0
   sqflite_common_ffi: ^2.3.0+4
   sqflite_sqlcipher: ^2.2.1
+  sqlite3: ^2.3.0
   swipe_to_action: ^0.2.0
   tor_detector_web: ^1.1.0
   uni_links: ^0.5.1
